% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rm_0s_by_col}
\alias{rm_0s_by_col}
\title{To remove columns of a data frame containing only 0's.}
\usage{
rm_0s_by_col(x)
}
\arguments{
\item{x}{Input data frame}
}
\description{
To remove columns of a data frame containing only 0's.
}
