% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rm_0s_by_row}
\alias{rm_0s_by_row}
\title{To remove rows of a data frame containing only 0's.}
\usage{
rm_0s_by_row(x)
}
\arguments{
\item{x}{Input data frame}
}
\description{
To remove rows of a data frame containing only 0's.
}
